Add new method run_cmd with raisesysexit argument.
In order to run commands which could fail in normal circumstances
use run_cmd with raisesysexit=False. This function will not raise
a SystemExit exception and returns a tuple of return code and command
output which used to evaluate the status of the executed command.
safe_cmd calls run_cmd with raisesysexit=True which will raise a
SystemExit exception if the return code is non-zero (previous behaviour).