fix problem with outdated git refs in cache
With the --shared option the git object store and the refs
were taken from the local cache directory. This lead to outdated
repository as the locally cached branches have not been updated.
Now we use the refs from the original git repository and the object
store from the local cache. Updating the cache via fetch is done
as before.