tests: add helper for GPG key-pair setup
The GPG key-pair is created in a "gpg" subdirectory under the parent
test dir. Once initialised, the GNUPGHOME environment variable is set
for all safe_run() command invocations, ensuring that it can be used
for signing and verification.

Both Git and Bazaar support signing and verification, so add this
functionality to the common Fixtures class.

Signed-off-by: David Disseldorp <ddiss@suse.de>