[bugfix] Decoupled self.scm from class name in TarSCM/scm/*
* fixes https://github.com/openSUSE/obs-service-tar_scm/issues/175

Without this patch the class name is used as self.scm in TarSCM.scm.*
classes. This value is also used to strip off the scm extension from the
url to get the repodir

Example:

https://github.com/openSUSE/obs-service-tar_scm.git -> obs-service-tar_scm

This broke after renaming scm classes to camel case for better linting
results.

In this patch, self.scm is changed to an class attribute
and decoupled from the class name.