Add new archive option for Debian: git-buildpackage
Debian + git developers workflow never include the .dsc file in the
repository, as it is the output of a build, not the input.
This makes it awkward and convoluted to use tar_scm with Debian git
repositories, and just unfeasable with obs_scm.

The most common workflow for Debian developers involves a tool called
git-buildpackage, that can generate the Debian source artefacts
(dsc file, orig.tar.gz and if non-native debian.tar.gz or diff.gz)
from a git repository.

Add a new class in the archive module that calls git-buildpackage as
an alternative to creating a tar archive or an obscpio.

This makes it possible to take advantage of the shared git clones
cache that obs_scm uses, saving a lot of network I/O.

The code is adapted from obs-service-git_buildpackage:
https://github.com/jblunck/obs-service-git_buildpackage