Version date cleanup: don't strip if version is not a date
version_iso_cleanup unconditionally strips : and - characters from the
version string when attempting to sanitize to ISO format. But this is
done regardless of whether the version is a date or not, which breaks
Debian packages, that use - as a separator for the upstream vs
downstream revisions.
Remove the characters only if the version matches the date format
regex.