add an option to choose what to replace non iso characters with
Some upstream repos use annoying version format like: 10.2-63.
When running through the service, : and - are deleted which generates
unusable/misleading version strings (10.263).

This patches adds an option (iso-cleanup-string) which allows to choose which charater/string
to use to replace the non iso ones.
Default value is '' which keeps the legacy behaviour.

For example
10.2-63 with iso-cleanup-string='.' will generate 10.2.63

Signed-off-by: Nicolas Morey-Chaisemartin <NMoreyChaisemartin@suse.com>