Add include_topdir option for the tar service
This commit adds include_topdir client option for the tar service.
Concretely it enables the user to choose if the top directory of the
content that has been required to tar should or not be included in the
resulting tarball. This is something revelant in order to allow OBS
services to fetch KIWI projects from some SCM.

Fixes #190