changed PREFIX in Gnumakefile to /usr
Installing under /usr/local seems quite useless for me because it makes
debugging harder as I always start with the command line given in bs_service
logfile. Most of the time I do at first a "make install" and wonder why my
changes are not used.

This change is only for convenience