Check that parent_tag is in an ancestor of the checked out branch
`git rev-list A..B` returns the set of commits present in B but not
A. Only if A is an ancestor of B will the cardinality of the set be
the same as the forward-running offset from A to B.

When parent-tag is implicit, the tag `git-describe` chooses is
already known to be an ancestor. Only when this is explicitly
overridden with a different parent-tag in the _service file, it will
be possible for the chosen parent-tag value not to refer to a commit
that is a valid ancestor.