Version cleanup: don't strip hyphen when building Debian packages
version_iso_cleanup unconditionally strips : and - characters from the
version string when attempting to sanitize to ISO format.
This is necessary because they are illegal characters in RPM format.

But they are legal for Debian versions: - splits the upstream versus
downstream versions, and : is used for epochs, and more importantly
stripping them breaks package builds.
Do not strip them if using the new obs_gbp service as a workaround
that does not have an impact on RPM builds.

Add a simple unit test for this function.