improve handling of corrupt git cache directories
Without this patch, a corrupt cache directory will never get cleaned and
result in broken services until admins runs a cleanup job on the server
side.

With
https://github.com/openSUSE/open-build-service/commit/5d49a2897f597e3db6ca4efc2572bd1a6564e4df
we added a new ENV var "OBS_SERVICE_DAEMON" in bs_service, so we can ensure
that we are running in server context. If server context is detected, broken i
cache will be removed automatically.