raise exception if no changesauthor given
This patch checks if a changesauthor is either

* given on cli (or _service file)
* or set in environment as VC_MAILADDR (added recently to osc)

If not running in OBS_SERVICE_DAEMON=1 (set by bs_service), we want services
with enabled "changesgenerate" to fail if no "changesauthor" is set. Package
maintainers are supposed to configure this option properly before submitting
to openSUSE:Factory or any other openSUSE/SUSE distribution.

fixing testcases with "--changesauthor"