Prefer UTF-8 locale as output format for changes
The old code had a kinda inverted logic ...

1. we guessed the encoding from the new changes entry
2. encoded the lines with that
3. tried to reencode the old messages also to that encodeing

so if the new messages would match ascii, it would try to reencode
everything as ascii. this can not work.

mls said multiple times that rpm wants to have everything as UTF-8
strings. so why not have everything as UTF-8 as early as possible?

During a discussion with Marcus Huewe, we came to the idea that we could
add a "osc vc --recode" command that would take an existing changes file
and convert each entry to UTF-8. Each entry as the encoding can differ
from entry to entry. In theory even from word to word in an entry. But
lets be optimistic for now.

For the same reason I propose we set the default locale to en_US.UTF-8.
C.UTF-8 is not an official locale, so en_US.UTF-8 seems to be a good
compromise.

Force cmd run helper to return UTF-8 strings

This saves us a lot of juggling with encoding later on.
Tested in python2 and python3 but only with git.

Force en_US.UTF-8 locale for the testsuite

As we force the en_US.UTF-8 locale in cli.py as well, but it seems the
testsuite skips that

Make sure the travis build has the locale built

just set LC_ALL. it should take care of LC_CTYPE aswell

Remove runtime deps on six and chardet