Fix the logic to pick the locale package on Fedora
and RHEL/Centos

the old code in the spec file was

0%{?fedora_version} < 27 || 0%{?rhel_version} < 8 || 0%{?centos_version} < 8

no matter on which distro you build one of those was always false
and that means we never pulled the package for newer distros.

The newer logic fixes this. Also in git this conditional was missing,
while in openSUSE:Tools it was present. This commit also fixes this.