Override LANGUAGE and LC_ALL to force commands run to use english
Having LANGUAGE=es in the environment was making `osc service disabledrun`
fail because the `No local changes to save\n` text couldn't be parsed by
scm/git.py since git was running in spanish as LANGUAGE has precedence over
LANG, so we now set it too to force using english.

Just to be sure we don't have a similar issue in the future with
different encodings being used because LC_ALL has precedence over LANG,
we also override LC_ALL so we're safe.