Don`t break testsuite if cwd contains colons
So far we created a directory which contains symbolic links named like
the scm to be tests (e.g. git/svn/bzr/hg etc.) in a subdiretory under
"__file__/tmp/..." which normally is "os.getcwd()/tests/tmp/...". This
path was exported by "os.environ['PATH'] = ....".

This leads to a broken test suite if the current working directory contains
colons, as colon is the separator sign in the PATH env variable.

This patch create a tempdir under "/tmp" which now contains the symbolic links
to the scm-wrapper script to be independent from the current working directory
when executing the test suite.