Append '_service' to repository directory
A package may contain a file with the same name as the package.
In this case, this file cannot be extracted, as the repository
directory is already there.
Append '_service' to the name of the directory to clone to to
avoid this trap.
Save the name of this directory for later use in obsinfo.

Signed-off-by: Egbert Eich <eich@suse.com>