dist: Convert to multibuild; run test suite in a separate flavor
Allows the obs-service-tar_scm package to be used before running
the test suite. This is especially interesting in the case of
distro-bootstrapping, as e.g. python3 does not get all its
dependencies before tar_scm is available.

For Stagings, we basically already have the --without-obs_scm_testing
in place, but that only works partially (due to BuildRequires python3).