Disable code which corrupts versions
A _service file which intends to use @PARENT_TAG@ for the package
version, and which uses versionrewrite-pattern to process the tag
further, will be unable to process the full tag name to extract the
desired parts of it via versionrewrite-replacement. The reason is the
forced removal of all hyphens.

It is up to the author of the _service file how the final package
version is assembled. versionrewrite-pattern and
versionrewrite-replacement give control how this needs to be done.

Adjust detect_version to keep hyphens if the _service file uses
versionrewrite-pattern, which should have been done along with the
changes that introduced versionrewrite in commit
960d0cebad3bbf86102991eb38c542bbf5a16df4.

Fixes issue #367

Signed-off-by: Olaf Hering <olaf@aepfle.de>