add '--' to git log command if file/dir equal revision exists
Fixes #412

Without this patch obs-service-tar_scm fails with an error message like

```
fatal: ambiguous argument ...
```

if a file/directory named like a tag/branch/revision exists
in the git top level dir.

This patch adds '--' to the list of arguments of the "git log" command if a
revision is set and a file <clone_dir>/<revision> exists to avoid errors like
mentioned above.