Change date format from short to %Y%m%d.
Since commit 49bb443 ("don`t sanitize version if versionrewrite_pattern
is set") the dashes from the short date format get into the version
string when versionrewrite_pattern is specified.

Change the date format so that no dashes are included in the date to
start with so that we don't need to rely on the sanitization to remove
them again.

Fixes: #419
Fixes: 49bb443 ("don`t sanitize version if versionrewrite_pattern is set")
Signed-off-by: Michal Suchanek <msuchanek@suse.de>