Fix Change Tar Archive to use gzip compression
This commit modifies the Tar class in BaseArchive to output gzip-compressed tar files instead of uncompressed tar files. This change improves the efficiency of the archive process by reducing the size of the output files.

This change is backward compatible, as tar and gzip are standard on most systems and gzip decompression is automatically handled by tar when extracting the archive.