Feat add gzip compression option for tarball creation
This commit introduces the option to create a gzip compressed tarball 
when the 'tar.gz' extension is selected. The open mode for the tarfile is 
determined based on the chosen extension.

If the extension is 'tar.gz', a gzip compressed tarball will be created. 
Otherwise, a regular uncompressed tarball will be created.

This allows for greater flexibility when creating tarballs, 
without breaking existing setups that do not use gzip compression.