disable compile python in debian package
Lintian complains:

E: obs-service-tar-scm: package-installs-python-pycache-dir
usr/lib/obs/service/TarSCM/__pycache__/

https://lintian.debathena.org/tags/package-installs-python-pycache-dir.html

as debian packages should not install compiled python; package install
hooks take care of that.

So add a variable to control this from the makefile, and disable it from
the debian package build.

DESTDIR and PREFIX do not need to be explictly set if just passing
additional options to the dh_auto_install instead of completely
overrriding it.