gbp: use --git-export
`--git-export` only works when `--git-export-dir` is set.

As the service `scm_object` has already cloned and setup the appropriate
revision in the working copy, and as the working copy of
`debian/changelog` is modified, export the working-copy using the gbp
special `WC` identifier so this is preserved.

This also fixes an issue with the `debian/changelog` revision
modification being lost if a deployment has global export hooks
configured in `/etc/git-buildpackage/gbp.conf` and export enabled.