Add --ignore-for-workflow parameter
When a _service contains multiple obs_scm entries, the OBS SCM
workflow branch_package will cause all of them to be updated to
the fork/commit of the workflow. But this is rarely what we want,
as for example in the case where the source code to build is in
one main repository (that should be overridden), but there are
also auxiliary repositories where the packaging recipes are
stored, and that should not be updated, as the references will
be invalid.

This is making it impossible to use branch_package on PRs for
the systemd upstream CI project:

https://build.opensuse.org/project/show/home:bluca:systemd

The new parameters allows me to configure the auxiliary sources
to be skipped by the branch_package update.