Use fnmatch again for excludes
According to the documentation, the value of the "excludes" parameter for
obs_scm is a glob. But since commit 140a127 ("fix excludes for obscpio") the
code treats this value as a regular expression, which leads to wrong matching
results and may lead to python crashes if the glob expression starts with "*".

Fixes: https://github.com/openSUSE/obs-service-tar_scm/issues/520
Fixes: 140a127 ("fix excludes for obscpio")